Check to see if document was selected

+++ main.py
@@ -2,6 +2,9 @@ from documentcloud.addon import SoftTimeOutAddOn
 
 class Clear(SoftTimeOutAddOn):
     def main(self):
+        if not self.documents:
+            self.set_message("Please select at least one document.")
+            return
         for document in self.get_documents():
             if document.status == 'error' or document.status == 'nofile':
                 document.delete()
